Updated. Need to correct %d, %i, %x and %X

+++ ft_printf/ft_printdec.c
@@ -6,13 +6,13 @@
 /*   By: dacaball <dacaball@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 16:24:57 by dacaball          #+#    #+#             */
-/*   Updated: 2023/10/05 16:45:01 by dacaball         ###   ########.fr       */
+/*   Updated: 2023/10/06 11:33:13 by dacaball         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_printdec(int decimal, int *i)
+void	ft_printdec(long int decimal, int *i)
 {	
 	if (decimal == -2147483648)
 	{
